+++ src/plugins/discourse/createGraph.js
@@ -1,211 +1,14 @@
 // @flow
 
-import {Graph, EdgeAddress, type Node, type Edge} from "../../core/graph";
-import {
-  type PostId,
-  type TopicId,
-  type Post,
-  type Topic,
-  type LikeAction,
-} from "./fetch";
+import {Graph, type Edge} from "../../core/graph";
+import {type PostId, type TopicId, type Post, type LikeAction} from "./fetch";
 import {type ReadRepository} from "./mirrorRepository";
 import {
-  authorsPostEdgeType,
-  authorsTopicEdgeType,
-  postRepliesEdgeType,
-  topicContainsPostEdgeType,
-  likesEdgeType,
-  createsLikeEdgeType,
-  referencesTopicEdgeType,
-  referencesUserEdgeType,
-  referencesPostEdgeType,
-} from "./declaration";
-import {userAddress, topicAddress, postAddress, likeAddress} from "./address";
-import {
   type DiscourseReference,
-  type DiscourseTopicReference,
-  type DiscourseUserReference,
   parseLinks,
   linksToReferences,
 } from "./references";
-
-export function userNode(serverUrl: string, username: string): Node {
-  const url = `${serverUrl}/u/${username}/`;
-  const description = `[@${username}](${url})`;
-  return {
-    address: userAddress(serverUrl, username),
-    description,
-    timestampMs: null,
-  };
-}
-
-export function topicNode(serverUrl: string, topic: Topic): Node {
-  const url = `${serverUrl}/t/${String(topic.id)}`;
-  const description = `[${topic.title}](${url})`;
-  const address = topicAddress(serverUrl, topic.id);
-  return {address, description, timestampMs: topic.timestampMs};
-}
-
-export function postNode(
-  serverUrl: string,
-  post: Post,
-  description: string
-): Node {
-  const address = postAddress(serverUrl, post.id);
-  return {timestampMs: post.timestampMs, address, description};
-}
-
-export function likeNode(
-  serverUrl: string,
-  like: LikeAction,
-  postDescription: string
-): Node {
-  const address = likeAddress(serverUrl, like);
-  const description = `❤️ by ${like.username} on ${postDescription}`;
-  return {timestampMs: like.timestampMs, address, description};
-}
-
-export function authorsTopicEdge(serverUrl: string, topic: Topic): Edge {
-  const address = EdgeAddress.append(
-    authorsTopicEdgeType.prefix,
-    serverUrl,
-    topic.authorUsername,
-    String(topic.id)
-  );
-  return {
-    address,
-    timestampMs: topic.timestampMs,
-    src: userAddress(serverUrl, topic.authorUsername),
-    dst: topicAddress(serverUrl, topic.id),
-  };
-}
-
-export function authorsPostEdge(serverUrl: string, post: Post): Edge {
-  const address = EdgeAddress.append(
-    authorsPostEdgeType.prefix,
-    serverUrl,
-    post.authorUsername,
-    String(post.id)
-  );
-  return {
-    address,
-    timestampMs: post.timestampMs,
-    src: userAddress(serverUrl, post.authorUsername),
-    dst: postAddress(serverUrl, post.id),
-  };
-}
-
-export function topicContainsPostEdge(serverUrl: string, post: Post): Edge {
-  const address = EdgeAddress.append(
-    topicContainsPostEdgeType.prefix,
-    serverUrl,
-    String(post.topicId),
-    String(post.id)
-  );
-  return {
-    address,
-    timestampMs: post.timestampMs,
-    src: topicAddress(serverUrl, post.topicId),
-    dst: postAddress(serverUrl, post.id),
-  };
-}
-
-export function postRepliesEdge(
-  serverUrl: string,
-  post: Post,
-  basePostId: PostId
-): Edge {
-  const address = EdgeAddress.append(
-    postRepliesEdgeType.prefix,
-    serverUrl,
-    String(post.id),
-    String(basePostId)
-  );
-  return {
-    address,
-    timestampMs: post.timestampMs,
-    src: postAddress(serverUrl, post.id),
-    dst: postAddress(serverUrl, basePostId),
-  };
-}
-
-export function createsLikeEdge(serverUrl: string, like: LikeAction): Edge {
-  const address = EdgeAddress.append(
-    createsLikeEdgeType.prefix,
-    serverUrl,
-    like.username,
-    String(like.postId)
-  );
-  return {
-    address,
-    timestampMs: like.timestampMs,
-    src: userAddress(serverUrl, like.username),
-    dst: likeAddress(serverUrl, like),
-  };
-}
-
-export function likesEdge(serverUrl: string, like: LikeAction): Edge {
-  const address = EdgeAddress.append(
-    likesEdgeType.prefix,
-    serverUrl,
-    like.username,
-    String(like.postId)
-  );
-  return {
-    address,
-    timestampMs: like.timestampMs,
-    src: likeAddress(serverUrl, like),
-    dst: postAddress(serverUrl, like.postId),
-  };
-}
-
-export function referencesTopicEdge(
-  serverUrl: string,
-  post: Post,
-  reference: DiscourseTopicReference
-): Edge {
-  const address = EdgeAddress.append(
-    referencesTopicEdgeType.prefix,
-    serverUrl,
-    String(post.id),
-    String(reference.topicId)
-  );
-  const src = postAddress(serverUrl, post.id);
-  const dst = topicAddress(serverUrl, reference.topicId);
-  return {src, dst, timestampMs: post.timestampMs, address};
-}
-
-export function referencesPostEdge(
-  serverUrl: string,
-  post: Post,
-  referredPostId: PostId
-): Edge {
-  const address = EdgeAddress.append(
-    referencesPostEdgeType.prefix,
-    serverUrl,
-    String(post.id),
-    String(referredPostId)
-  );
-  const src = postAddress(serverUrl, post.id);
-  const dst = postAddress(serverUrl, referredPostId);
-  return {src, dst, timestampMs: post.timestampMs, address};
-}
-
-export function referencesUserEdge(
-  serverUrl: string,
-  post: Post,
-  reference: DiscourseUserReference
-): Edge {
-  const address = EdgeAddress.append(
-    referencesUserEdgeType.prefix,
-    serverUrl,
-    String(post.id),
-    reference.username
-  );
-  const src = postAddress(serverUrl, post.id);
-  const dst = userAddress(serverUrl, reference.username);
-  return {src, dst, timestampMs: post.timestampMs, address};
-}
+import * as NE from "./nodesAndEdges";
 
 export function createGraph(serverUrl: string, data: ReadRepository): Graph {
   const gc = new _GraphCreator(serverUrl, data);
@@ -230,13 +33,13 @@ class _GraphCreator {
     this.postIdToDescription = new Map();
 
     for (const username of data.users()) {
-      this.graph.addNode(userNode(serverUrl, username));
+      this.graph.addNode(NE.userNode(serverUrl, username));
     }
 
     for (const topic of data.topics()) {
       this.topicIdToTitle.set(topic.id, topic.title);
-      this.graph.addNode(topicNode(serverUrl, topic));
-      this.graph.addEdge(authorsTopicEdge(serverUrl, topic));
+      this.graph.addNode(NE.topicNode(serverUrl, topic));
+      this.graph.addEdge(NE.authorsTopicEdge(serverUrl, topic));
     }
 
     for (const post of data.posts()) {
@@ -256,9 +59,9 @@ class _GraphCreator {
   }
 
   addPost(post: Post, description: string) {
-    this.graph.addNode(postNode(this.serverUrl, post, description));
-    this.graph.addEdge(authorsPostEdge(this.serverUrl, post));
-    this.graph.addEdge(topicContainsPostEdge(this.serverUrl, post));
+    this.graph.addNode(NE.postNode(this.serverUrl, post, description));
+    this.graph.addEdge(NE.authorsPostEdge(this.serverUrl, post));
+    this.graph.addEdge(NE.topicContainsPostEdge(this.serverUrl, post));
     this.maybeAddPostRepliesEdge(post);
 
     const discourseReferences = linksToReferences(
@@ -275,9 +78,9 @@ class _GraphCreator {
   addLike(like: LikeAction) {
     const postDescription =
       this.postIdToDescription.get(like.postId) || "[unknown post]";
-    this.graph.addNode(likeNode(this.serverUrl, like, postDescription));
-    this.graph.addEdge(likesEdge(this.serverUrl, like));
-    this.graph.addEdge(createsLikeEdge(this.serverUrl, like));
+    this.graph.addNode(NE.likeNode(this.serverUrl, like, postDescription));
+    this.graph.addEdge(NE.likesEdge(this.serverUrl, like));
+    this.graph.addEdge(NE.createsLikeEdge(this.serverUrl, like));
   }
 
   /**
@@ -300,7 +103,9 @@ class _GraphCreator {
         replyToPostIndex
       );
       if (basePostId != null) {
-        this.graph.addEdge(postRepliesEdge(this.serverUrl, post, basePostId));
+        this.graph.addEdge(
+          NE.postRepliesEdge(this.serverUrl, post, basePostId)
+        );
       }
     }
   }
@@ -316,7 +121,7 @@ class _GraphCreator {
     }
     switch (reference.type) {
       case "TOPIC": {
-        return referencesTopicEdge(this.serverUrl, post, reference);
+        return NE.referencesTopicEdge(this.serverUrl, post, reference);
       }
       case "POST": {
         const referredPostId = this.data.findPostInTopic(
@@ -327,10 +132,10 @@ class _GraphCreator {
           // Maybe a bad link, or the post or topic was deleted.
           return null;
         }
-        return referencesPostEdge(this.serverUrl, post, referredPostId);
+        return NE.referencesPostEdge(this.serverUrl, post, referredPostId);
       }
       case "USER": {
-        return referencesUserEdge(this.serverUrl, post, reference);
+        return NE.referencesUserEdge(this.serverUrl, post, reference);
       }
       default: {
         throw new Error(
