+++ src/plugins/discourse/createGraph.test.js
@@ -4,22 +4,11 @@ import sortBy from "../../util/sortBy";
 import * as NullUtil from "../../util/null";
 import type {ReadRepository} from "./mirrorRepository";
 import type {Topic, Post, PostId, TopicId, LikeAction} from "./fetch";
-import {NodeAddress, EdgeAddress, type Node, type Edge} from "../../core/graph";
-import {
-  createGraph,
-  userNode,
-  topicNode,
-  postNode,
-  likeNode,
-  authorsTopicEdge,
-  authorsPostEdge,
-  topicContainsPostEdge,
-  postRepliesEdge,
-  likesEdge,
-  createsLikeEdge,
-} from "./createGraph";
+import {EdgeAddress, type Node, type Edge} from "../../core/graph";
+import {createGraph} from "./createGraph";
+import * as NE from "./nodesAndEdges";
 
-import {userAddress, postAddress, topicAddress, likeAddress} from "./address";
+import {userAddress, postAddress, topicAddress} from "./address";
 
 import {
   userNodeType,
@@ -151,83 +140,6 @@ describe("plugins/discourse/createGraph", () => {
   }
 
   describe("nodes are constructed correctly", () => {
-    it("for users", () => {
-      const {url} = example();
-      const node = userNode(url, "decentralion");
-      expect(node.description).toMatchInlineSnapshot(
-        `"[@decentralion](https://url.com/u/decentralion/)"`
-      );
-      expect(node.timestampMs).toEqual(null);
-      expect(NodeAddress.toParts(node.address)).toMatchInlineSnapshot(`
-                                Array [
-                                  "sourcecred",
-                                  "discourse",
-                                  "user",
-                                  "https://url.com",
-                                  "decentralion",
-                                ]
-                        `);
-    });
-
-    it("for topics", () => {
-      const {url, topic} = example();
-      const node = topicNode(url, topic);
-      expect(node.description).toMatchInlineSnapshot(
-        `"[first topic](https://url.com/t/1)"`
-      );
-      expect(node.timestampMs).toEqual(topic.timestampMs);
-      expect(NodeAddress.toParts(node.address)).toMatchInlineSnapshot(`
-                                Array [
-                                  "sourcecred",
-                                  "discourse",
-                                  "topic",
-                                  "https://url.com",
-                                  "1",
-                                ]
-                        `);
-    });
-
-    it("for posts", () => {
-      const {url, posts} = example();
-      const description = "[#2 on first topic](https://url.com/t/1/2)";
-      const node = postNode(url, posts[1], description);
-      expect(node.description).toEqual(description);
-      expect(node.timestampMs).toEqual(posts[1].timestampMs);
-      expect(NodeAddress.toParts(node.address)).toMatchInlineSnapshot(`
-                                Array [
-                                  "sourcecred",
-                                  "discourse",
-                                  "post",
-                                  "https://url.com",
-                                  "2",
-                                ]
-                        `);
-    });
-
-    it("for likes", () => {
-      const {url, likes, posts, graph} = example();
-      const like = likes[0];
-      const post = posts[1];
-      expect(like.postId).toEqual(post.id);
-      const postDescription = `[#2 on first topic](https://url.com/t/1/2)`;
-      const node = likeNode(url, like, postDescription);
-      expect(node.description).toMatchInlineSnapshot(
-        `"❤️ by mzargham on [#2 on first topic](https://url.com/t/1/2)"`
-      );
-      expect(node.timestampMs).toEqual(like.timestampMs);
-      expect(NodeAddress.toParts(node.address)).toMatchInlineSnapshot(`
-        Array [
-          "sourcecred",
-          "discourse",
-          "like",
-          "https://url.com",
-          "mzargham",
-          "2",
-        ]
-      `);
-      expect(graph.node(node.address)).toEqual(node);
-    });
-
     it("gives an [unknown post] description for likes without a matching post", () => {
       const {likes} = example();
       const like = likes[0];
@@ -235,7 +147,7 @@ describe("plugins/discourse/createGraph", () => {
       const url = "https://foo";
       const graph = createGraph(url, data);
       const actual = Array.from(graph.nodes())[0];
-      const expected = likeNode(url, like, "[unknown post]");
+      const expected = NE.likeNode(url, like, "[unknown post]");
       expect(actual).toEqual(expected);
     });
 
@@ -257,133 +169,8 @@ describe("plugins/discourse/createGraph", () => {
       )}`;
       const expectedDescription = `[#${post.indexWithinTopic} on [unknown topic]](${postUrl})`;
       const actual = Array.from(graph.nodes({prefix: postNodeType.prefix}))[0];
-      const expected = postNode(url, post, expectedDescription);
+      const expected = NE.postNode(url, post, expectedDescription);
       expect(actual).toEqual(expected);
-    });
-  });
-
-  describe("edges are constructed correctly", () => {
-    it("for authorsTopic", () => {
-      const {url, topic} = example();
-      const expectedSrc = userNode(url, topic.authorUsername).address;
-      const expectedDst = topicNode(url, topic).address;
-      const edge = authorsTopicEdge(url, topic);
-      expect(edge.src).toEqual(expectedSrc);
-      expect(edge.dst).toEqual(expectedDst);
-      expect(edge.timestampMs).toEqual(topic.timestampMs);
-      expect(EdgeAddress.toParts(edge.address)).toMatchInlineSnapshot(`
-                        Array [
-                          "sourcecred",
-                          "discourse",
-                          "authors",
-                          "topic",
-                          "https://url.com",
-                          "decentralion",
-                          "1",
-                        ]
-                  `);
-    });
-    it("for authorsPost", () => {
-      const {url, posts, topic} = example();
-      const post = posts[1];
-      const expectedSrc = userNode(url, post.authorUsername).address;
-      const expectedDst = postNode(url, post, topic.title).address;
-      const edge = authorsPostEdge(url, post);
-      expect(edge.src).toEqual(expectedSrc);
-      expect(edge.dst).toEqual(expectedDst);
-      expect(edge.timestampMs).toEqual(post.timestampMs);
-      expect(EdgeAddress.toParts(edge.address)).toMatchInlineSnapshot(`
-                        Array [
-                          "sourcecred",
-                          "discourse",
-                          "authors",
-                          "post",
-                          "https://url.com",
-                          "wchargin",
-                          "2",
-                        ]
-                  `);
-    });
-    it("for topicContainsPost", () => {
-      const {url, posts, topic} = example();
-      const post = posts[1];
-      const expectedSrc = topicNode(url, topic).address;
-      const expectedDst = postNode(url, post, topic.title).address;
-      const edge = topicContainsPostEdge(url, post);
-      expect(edge.src).toEqual(expectedSrc);
-      expect(edge.dst).toEqual(expectedDst);
-      expect(edge.timestampMs).toEqual(post.timestampMs);
-      expect(EdgeAddress.toParts(edge.address)).toMatchInlineSnapshot(`
-                        Array [
-                          "sourcecred",
-                          "discourse",
-                          "topicContainsPost",
-                          "https://url.com",
-                          "1",
-                          "2",
-                        ]
-                  `);
-    });
-    it("for postReplies", () => {
-      const {url, posts, topic} = example();
-      const post = posts[2];
-      const basePost = posts[1];
-      const expectedSrc = postNode(url, post, topic.title).address;
-      const expectedDst = postNode(url, basePost, topic.title).address;
-      const edge = postRepliesEdge(url, post, basePost.id);
-      expect(edge.src).toEqual(expectedSrc);
-      expect(edge.dst).toEqual(expectedDst);
-      expect(edge.timestampMs).toEqual(post.timestampMs);
-      expect(EdgeAddress.toParts(edge.address)).toMatchInlineSnapshot(`
-                        Array [
-                          "sourcecred",
-                          "discourse",
-                          "replyTo",
-                          "https://url.com",
-                          "3",
-                          "2",
-                        ]
-                  `);
-    });
-    it("for likes", () => {
-      const {url, likes} = example();
-      const like = likes[0];
-      const expectedSrc = likeAddress(url, like);
-      const expectedDst = postAddress(url, like.postId);
-      const edge = likesEdge(url, like);
-      expect(edge.src).toEqual(expectedSrc);
-      expect(edge.dst).toEqual(expectedDst);
-      expect(edge.timestampMs).toEqual(like.timestampMs);
-      expect(EdgeAddress.toParts(edge.address)).toMatchInlineSnapshot(`
-        Array [
-          "sourcecred",
-          "discourse",
-          "likes",
-          "https://url.com",
-          "mzargham",
-          "2",
-        ]
-      `);
-    });
-    it("for createsLike", () => {
-      const {url, likes} = example();
-      const like = likes[0];
-      const expectedSrc = userAddress(url, like.username);
-      const expectedDst = likeAddress(url, like);
-      const edge = createsLikeEdge(url, like);
-      expect(edge.src).toEqual(expectedSrc);
-      expect(edge.dst).toEqual(expectedDst);
-      expect(edge.timestampMs).toEqual(like.timestampMs);
-      expect(EdgeAddress.toParts(edge.address)).toMatchInlineSnapshot(`
-        Array [
-          "sourcecred",
-          "discourse",
-          "createsLike",
-          "https://url.com",
-          "mzargham",
-          "2",
-        ]
-      `);
     });
   });
 
@@ -398,12 +185,12 @@ describe("plugins/discourse/createGraph", () => {
     it("for users", () => {
       const {url} = example();
       const usernames = ["decentralion", "wchargin", "mzargham"];
-      const expected = usernames.map((x) => userNode(url, x));
+      const expected = usernames.map((x) => NE.userNode(url, x));
       expectNodesOfType(expected, userNodeType);
     });
     it("for topics", () => {
       const {url, topic} = example();
-      const expected = [topicNode(url, topic)];
+      const expected = [NE.topicNode(url, topic)];
       expectNodesOfType(expected, topicNodeType);
     });
     it("for posts", () => {
@@ -413,7 +200,7 @@ describe("plugins/discourse/createGraph", () => {
           x.indexWithinTopic
         )}`;
         const description = `[#${x.indexWithinTopic} on ${topic.title}](${postUrl})`;
-        return postNode(url, x, description);
+        return NE.postNode(url, x, description);
       });
       expectNodesOfType(expected, postNodeType);
     });
@@ -428,7 +215,7 @@ describe("plugins/discourse/createGraph", () => {
         postIdToDescription.set(post.id, description);
       }
       const expected = likes.map((x) =>
-        likeNode(url, x, NullUtil.get(postIdToDescription.get(x.postId)))
+        NE.likeNode(url, x, NullUtil.get(postIdToDescription.get(x.postId)))
       );
       expectNodesOfType(expected, likeNodeType);
     });
@@ -446,36 +233,36 @@ describe("plugins/discourse/createGraph", () => {
     }
     it("authorsTopic edges", () => {
       const {url, topic} = example();
-      const topicEdge = authorsTopicEdge(url, topic);
+      const topicEdge = NE.authorsTopicEdge(url, topic);
       expectEdgesOfType([topicEdge], authorsTopicEdgeType);
     });
     it("authorsPost edges", () => {
       const {url, posts} = example();
-      const postEdges = posts.map((p) => authorsPostEdge(url, p));
+      const postEdges = posts.map((p) => NE.authorsPostEdge(url, p));
       expectEdgesOfType(postEdges, authorsPostEdgeType);
     });
     it("topicContainsPost edges", () => {
       const {url, posts} = example();
-      const edges = posts.map((p) => topicContainsPostEdge(url, p));
+      const edges = posts.map((p) => NE.topicContainsPostEdge(url, p));
       expectEdgesOfType(edges, topicContainsPostEdgeType);
     });
     it("postReplies edges", () => {
       const {url, posts} = example();
       const [post1, post2, post3] = posts;
       const edges = [
-        postRepliesEdge(url, post2, post1.id),
-        postRepliesEdge(url, post3, post2.id),
+        NE.postRepliesEdge(url, post2, post1.id),
+        NE.postRepliesEdge(url, post3, post2.id),
       ];
       expectEdgesOfType(edges, postRepliesEdgeType);
     });
     it("likes edges", () => {
       const {url, likes} = example();
-      const edges = likes.map((l) => likesEdge(url, l));
+      const edges = likes.map((l) => NE.likesEdge(url, l));
       expectEdgesOfType(edges, likesEdgeType);
     });
     it("createsLike edges", () => {
       const {url, likes} = example();
-      const edges = likes.map((l) => createsLikeEdge(url, l));
+      const edges = likes.map((l) => NE.createsLikeEdge(url, l));
       expectEdgesOfType(edges, createsLikeEdgeType);
     });
     it("references post edges", () => {
